+++ holoviews/element/geom.py
@@ -106,16 +106,17 @@ class Segments(Geometry):
         coordinates in 2D space.""")
 
 
-class Rects(Geometry):
+class Boxes(Geometry):
     """
-    Rects represent a collection of rectangles in 2D space.
+    Boxes represent a collection of axis-aligned rectangles in 2D space.
     """
 
-    group = param.String(default='Rects', constant=True)
+    group = param.String(default='Boxes', constant=True)
 
     kdims = param.List(default=[Dimension('x0'), Dimension('y0'),
                                 Dimension('x1'), Dimension('y1')],
                        bounds=(4, 4), constant=True, doc="""
-        Segments represent lines given by x- and y-
-        coordinates in 2D space.""")
+        The key dimensions of the Boxes element represent the
+        bottom-left (x0, y0) and top right (x1, y1) coordinates
+        of each box.""")
