+++ holoviews/plotting/bokeh/__init__.py
@@ -13,7 +13,7 @@ from ...element import (Curve, Points, Scatter, Image, Raster, Path,
                         Table, ItemTable, Area, HSV, QuadMesh, VectorField,
                         Graph, Nodes, EdgePaths, Distribution, Bivariate,
                         TriMesh, Violin, Chord, Div, HexTiles, Labels, Sankey,
-                        Tiles, Segments, Slope, Rects)
+                        Tiles, Segments, Slope, Boxes)
 from ...core.options import Options, Cycle, Palette
 from ...core.util import LooseVersion, VersionError
 
@@ -36,11 +36,12 @@ from .callbacks import Callback # noqa (API import)
 from .element import OverlayPlot, ElementPlot
 from .chart import (PointPlot, CurvePlot, SpreadPlot, ErrorPlot, HistogramPlot,
                     SideHistogramPlot, BarPlot, SpikesPlot, SideSpikesPlot,
-                    AreaPlot, VectorFieldPlot, SegmentPlot)
+                    AreaPlot, VectorFieldPlot)
+from .geometry import SegmentPlot, BoxesPlot
 from .graphs import GraphPlot, NodePlot, TriMeshPlot, ChordPlot
 from .heatmap import HeatMapPlot, RadialHeatMapPlot
 from .hex_tiles import HexTilesPlot
-from .path import PathPlot, PolygonPlot, ContourPlot, RectPlot
+from .path import PathPlot, PolygonPlot, ContourPlot
 from .plot import GridPlot, LayoutPlot, AdjointLayoutPlot
 from .raster import RasterPlot, RGBPlot, HSVPlot, QuadMeshPlot
 from .renderer import BokehRenderer
@@ -66,7 +67,6 @@ associations = {Overlay: OverlayPlot,
 
                 # Charts
                 Curve: CurvePlot,
-                Segments: SegmentPlot,
                 Bars: BarPlot,
                 Points: PointPlot,
                 Scatter: PointPlot,
@@ -96,7 +96,10 @@ associations = {Overlay: OverlayPlot,
                 Bounds:   PathPlot,
                 Ellipse:  PathPlot,
                 Polygons: PolygonPlot,
-                Rects:    RectPlot,
+
+                # Geometry
+                Boxes:    BoxesPlot,
+                Segments: SegmentPlot,
 
                 # Annotations
                 HLine: LineAnnotationPlot,
