+++ holoviews/plotting/bokeh/chart.py
@@ -4,13 +4,14 @@ from collections import defaultdict
 
 import numpy as np
 import param
+
 from bokeh.models import CategoricalColorMapper, CustomJS, Whisker, Range1d
 from bokeh.models.tools import BoxSelectTool
 from bokeh.transform import jitter
 
 from ...plotting.bokeh.selection import BokehOverlaySelectionDisplay
 from ...selection import NoOpSelectionDisplay
-from ...core.data import Dataset, Dimension
+from ...core.data import Dataset
 from ...core.dimension import dimension_name
 from ...core.util import (
     OrderedDict, max_range, basestring, dimension_sanitizer,
@@ -18,6 +19,7 @@ from ...core.util import (
 from ...element import Bars
 from ...operation import interpolate_curve
 from ...util.transform import dim
+from ..mixins import AreaMixin, SpikesMixin
 from ..util import compute_sizes, get_min_distance, get_axis_padding
 from .element import ElementPlot, ColorbarPlot, LegendPlot
 from .styles import (expand_batched_style, line_properties, fill_properties,
@@ -615,24 +617,9 @@ class SpreadPlot(ElementPlot):
 
 
 
-class AreaPlot(SpreadPlot):
+class AreaPlot(AreaMixin, SpreadPlot):
 
     _stream_data = False # Plot does not support streaming data
-
-    def get_extents(self, element, ranges, range_type='combined'):
-        vdims = element.vdims[:2]
-        vdim = vdims[0].name
-        if len(vdims) > 1:
-            new_range = {}
-            for r in ranges[vdim]:
-                new_range[r] = max_range([ranges[vd.name][r] for vd in vdims])
-            ranges[vdim] = new_range
-        else:
-            s0, s1 = ranges[vdim]['soft']
-            s0 = min(s0, 0) if isfinite(s0) else 0
-            s1 = max(s1, 0) if isfinite(s1) else 0
-            ranges[vdim]['soft'] = (s0, s1)
-        return super(AreaPlot, self).get_extents(element, ranges, range_type)
 
     def get_data(self, element, ranges, style):
         mapping = dict(x='x', y='y')
@@ -653,7 +640,7 @@ class AreaPlot(SpreadPlot):
 
 
 
-class SpikesPlot(ColorbarPlot):
+class SpikesPlot(SpikesMixin, ColorbarPlot):
 
     spike_length = param.Number(default=0.5, doc="""
       The length of each spike if Spikes object is one dimensional.""")
@@ -680,41 +667,6 @@ class SpikesPlot(ColorbarPlot):
         if 'spike_length' in self.lookup_options(element, 'plot').options:
             return  [element.dimensions()[0], None, None]
         return super(SpikesPlot, self)._get_axis_dims(element)
-
-    def get_extents(self, element, ranges, range_type='combined'):
-        opts = self.lookup_options(element, 'plot').options
-        if len(element.dimensions()) > 1 and 'spike_length' not in opts:
-            ydim = element.get_dimension(1)
-            s0, s1 = ranges[ydim.name]['soft']
-            s0 = min(s0, 0) if isfinite(s0) else 0
-            s1 = max(s1, 0) if isfinite(s1) else 0
-            ranges[ydim.name]['soft'] = (s0, s1)
-        proxy_dim = None
-        if 'spike_length' in opts:
-            proxy_dim = Dimension('proxy_dim')
-            proxy_range = (self.position, self.position + opts['spike_length'])
-            ranges['proxy_dim'] = {'data':    proxy_range,
-                                  'hard':     (np.nan, np.nan),
-                                  'soft':     (np.nan, np.nan),
-                                  'combined': proxy_range}
-        l, b, r, t = super(SpikesPlot, self).get_extents(element, ranges, range_type,
-                                                         ydim=proxy_dim)
-        if len(element.dimensions()) == 1 and range_type != 'hard':
-            if self.batched:
-                bs, ts = [], []
-                # Iterate over current NdOverlay and compute extents
-                # from position and length plot options
-                frame = self.current_frame or self.hmap.last
-                for el in frame.values():
-                    opts = self.lookup_options(el, 'plot').options
-                    pos = opts.get('position', self.position)
-                    length = opts.get('spike_length', self.spike_length)
-                    bs.append(pos)
-                    ts.append(pos+length)
-                b, t = (np.nanmin(bs), np.nanmax(ts))
-            else:
-                b, t = self.position, self.position+self.spike_length
-        return l, b, r, t
 
     def get_data(self, element, ranges, style):
         dims = element.dimensions()
@@ -1113,53 +1065,3 @@ class BarPlot(ColorbarPlot, LegendPlot):
                             'right': mapping.pop('top'), 'height': mapping.pop('width')})
 
         return sanitized_data, mapping, style
-
-class SegmentPlot(ColorbarPlot):
-    """
-    Segments are lines in 2D space where each two each dimensions specify a
-    (x, y) node of the line.
-    """
-    style_opts = line_properties + ['cmap']
-
-    _nonvectorized_styles = ['cmap']
-
-    _plot_methods = dict(single='segment')
-
-    def get_data(self, element, ranges, style):
-        # Get [x0, y0, x1, y1]
-        x0idx, y0idx, x1idx, y1idx = (
-            (1, 0, 3, 2) if self.invert_axes else (0, 1, 2, 3)
-        )
-
-        # Compute segments
-        x0s, y0s, x1s, y1s = (
-            element.dimension_values(x0idx),
-            element.dimension_values(y0idx),
-            element.dimension_values(x1idx),
-            element.dimension_values(y1idx)
-        )
-
-        data = {'x0': x0s, 'x1': x1s, 'y0': y0s, 'y1': y1s}
-        mapping = dict(x0='x0', x1='x1', y0='y0', y1='y1')
-        return (data, mapping, style)
-
-    def get_extents(self, element, ranges, range_type='combined'):
-        """
-        Use first two key dimensions to set names, and all four
-        to set the data range.
-        """
-        kdims = element.kdims
-        # loop over start and end points of segments
-        # simultaneously in each dimension
-        for kdim0, kdim1 in zip([kdims[i].name for i in range(2)],
-                                [kdims[i].name for i in range(2,4)]):
-            new_range = {}
-            for kdim in [kdim0, kdim1]:
-                # for good measure, update ranges for both start and end kdim
-                for r in ranges[kdim]:
-                    # combine (x0, x1) and (y0, y1) in range calculation
-                    new_range[r] = max_range([ranges[kd][r]
-                                              for kd in [kdim0, kdim1]])
-            ranges[kdim0] = new_range
-            ranges[kdim1] = new_range
-        return super(SegmentPlot, self).get_extents(element, ranges, range_type)
